perf: reduce memory usage for large-scale runs (500k+ agents)
Fix 1 - Period streaming (simulation_engine.py):
  - Replace unbounded all_visits accumulation with current_period buffer
  - current_period holds only the active eval window in RAM (eval_freq days)
  - After each evaluation, period is written to a temp parquet file and cleared
  - Remaining days flushed at end-of-run; all temp files read back and cleaned up
  - Peak visit RAM drops from total_days x rows/day to eval_freq x rows/day
  - Subcluster map built once upfront and attached before each flush

Fix 2 - float16 utility matrices (main.py, runner.py):
  - Change utility matrix dtype from float32 to float16
  - Halves matrix RAM: ~7.2 GB -> ~3.6 GB for 500k agents x 200 centres x 18 matrices
  - Safe for utility scores (bounded [0,1], no overflow risk at float16 range)

+++ simulation/main.py
@@ -163,7 +163,7 @@ class RetailABMApp:
                 cols = [c for c in df.columns if c.endswith(suf)]
                 if not cols: continue
                 
-                mat = df[cols].astype(np.float32)
+                mat = df[cols].astype(np.float16)
                 mat.columns = [_clean_rc_id(c[:-len(suf)]) for c in mat.columns]
                 
                 # Standardize index to string to avoid mixed-type duplication and row expansion
